Migrate from DocBook XML to ReStructuredText
As discussed in
https://github.com/sebastianbergmann/phpunit-documentation/issues/471.

In order to make contributing to the PHPUnit Documentation easier,
the existing DocBook XML files for the English PHPUnit 7.0
documentation have been converted to ReStructuredText files.

This not only makes writing the documentation easier, but it also
allows hosting it on readthedocs.org, which offers a very
straightforward way of handling translations.

Translations of the English version can now be kept in separate
repositories, allowing the translators to maintain them
themselves.